requirements.txt
django
pg8000
psycopg2-binary
requests
pandas
apscheduler
djangorestframework
sklearn
matplotlib
